+++ frontend/src/lib/usuarios/auth.ts
@@ -1,9 +1,13 @@
-import { UsuarioDtoOut } from "../type";
+import { UsuarioDtoOut, UsuarioForm} from "../type";
 
 const API_URL = process.env.NEXT_PUBLIC_API_URL;
 
 export async function fetchUsuarios(): Promise<UsuarioDtoOut[]> {
     const token = localStorage.getItem("token");
+    if (!token) {
+        window.location.href = "/login";
+        throw new Error("No hay sesión activa");
+    }
 
     const res = await fetch(`${API_URL}/usuarios`, {
         method: "GET",
@@ -28,9 +32,14 @@ export async function fetchUsuarios(): Promise<UsuarioDtoOut[]> {
     return res.json();
 }
 
-export async function deleteUsuario(id: number) {
+export async function deleteUsuario(idUsuario: number) {
     const token = localStorage.getItem("token");
-    const res = await fetch(`${API_URL}/usuarios/${id}`, {
+    if (!token) {
+        window.location.href = "/login";
+        throw new Error("No hay sesión activa");
+    }
+    
+    const res = await fetch(`${API_URL}/usuarios/${idUsuario}`, {
     method: "DELETE",
     headers: {
         "Authorization": `Bearer ${token}`,
@@ -55,4 +64,124 @@ export async function deleteUsuario(id: number) {
     }
 
     return true;
+}
+
+export async function postUsuario(formData: UsuarioForm): Promise<UsuarioDtoOut> {
+    const token = localStorage.getItem("token");
+    if (!token) {
+        window.location.href = "/login";
+        throw new Error("No hay sesión activa");
+    }
+
+    const body = {
+        nombre: formData.nombre,
+        apellido: formData.apellido,
+        contrasena: formData.contrasena,
+        cuil: formData.cuil,
+        mail: formData.mail,
+        rol: formData.rol
+    };
+
+    const res = await fetch(`${API_URL}/usuarios`, {
+        method: "POST",
+        headers: {
+        "Authorization": `Bearer ${token}`,
+        "Content-Type": "application/json",
+        },
+        body: JSON.stringify(body),
+    });
+
+    if (res.status === 401) {
+        localStorage.removeItem("token");
+        window.location.href = "/login";
+        throw new Error("Se perdió la sesión, redirigiendo a inicio de sesión");
+    }
+
+    if (res.status === 422) {
+        const err = await res.json();
+        throw new Error(err.detail || "Form mal formado");
+    }
+
+
+    if (!res.ok) {
+        const err = await res.json();
+        throw new Error(err.detail || "Error al crear el usuario");
+    }
+
+    return res.json();
+}
+
+export async function fetchUsuarioById(idUsuario: number): Promise<UsuarioDtoOut> {
+    const token = localStorage.getItem("token");
+    if (!token) {
+        window.location.href = "/login";
+        throw new Error("No hay sesión activa");
+    }
+
+    const res = await fetch(`${API_URL}/usuarios/${idUsuario}`, {
+        method: "GET",
+        headers: {
+        "Authorization": `Bearer ${token}`,
+        "Content-Type": "application/json",
+        },
+    });
+
+    if (res.status === 401) {
+        // limpiar sesión y redirigir
+        localStorage.removeItem("token")
+        window.location.href = "/login"
+        throw new Error("Se perdió la sesión, redirigiendo a inicio de sesión")
+    }
+
+    if (!res.ok) {
+        const err = await res.json();
+        throw new Error(err.detail || "Error al obtener el usuario");
+    }
+
+    return res.json();
+}
+
+export async function putUsuario(formData: UsuarioForm, idUsuario: number): Promise<UsuarioDtoOut> {
+    const token = localStorage.getItem("token");
+    if (!token) {
+        window.location.href = "/login";
+        throw new Error("No hay sesión activa");
+    }
+
+    const body = {
+        nombre: formData.nombre,
+        apellido: formData.apellido,
+        rol: formData.rol,
+        contrasena: formData.contrasena,
+        cuil: formData.cuil,
+        mail: formData.mail,
+    };
+
+    const res = await fetch(`${API_URL}/usuarios/${idUsuario}`, {
+        method: "PUT",
+        headers: {
+        "Authorization": `Bearer ${token}`,
+        "Content-Type": "application/json",
+        },
+        body: JSON.stringify(body),
+    });
+
+    if (res.status === 401) {
+        localStorage.removeItem("token");
+        window.location.href = "/login";
+        throw new Error("Se perdió la sesión, redirigiendo a inicio de sesión");
+    }
+
+    if (res.status === 422) {
+        const err = await res.json();
+        throw new Error(err.detail || "Form mal formado");
+    }
+
+
+    if (!res.ok) {
+        const err = await res.json();
+        throw new Error(err.detail || "Error al modificar el usuario");
+    }
+
+    return res.json();
 }
